Artist: Rework default artist creation
Don't INSERT OR IGNORE since this is only ran when we're creating the
database, or triggering a rescan. The previous artists shouldn't be
there anymore.

+++ src/MediaLibrary.cpp
@@ -2035,7 +2035,8 @@ bool MediaLibrary::forceRescan()
         Playlist::clearExternalPlaylistContent( this );
         parser::Task::removePlaylistContentTasks( this );
         parser::Task::resetParsing( this );
-        Artist::createDefaultArtists( getConn() );
+        if ( Artist::createDefaultArtists( getConn() ) == false )
+            return false;
         Thumbnail::deleteAll( this );
         Media::resetSubTypes( this );
         t->commit();
